+++ cynic-codegen/src/scalar_derive/mod.rs
@@ -39,5 +39,6 @@ pub fn scalar_derive_impl(input: ScalarDeriveInput) -> Result<TokenStream, syn::
             }
         }
 
+        ::cynic::impl_serializable_argument_for_scalar!(#ident);
     })
 }