+++ cynic/src/argument.rs
@@ -1,4 +1,4 @@
-use crate::{Scalar, SerializeError};
+use crate::SerializeError;
 
 pub struct Argument {
     pub(crate) name: String,
@@ -52,8 +52,33 @@ impl<T: SerializableArgument> SerializableArgument for Option<T> {
     }
 }
 
-impl<T: Scalar> SerializableArgument for T {
+impl<T: SerializableArgument> SerializableArgument for &T {
     fn serialize(&self) -> Result<serde_json::Value, SerializeError> {
-        self.encode()
+        (*self).serialize()
     }
 }
+
+impl SerializableArgument for &str {
+    fn serialize(&self) -> Result<serde_json::Value, SerializeError> {
+        Ok(serde_json::Value::String(self.to_string()))
+    }
+}
+
+/// Implements serializable argument for scalar types.
+#[macro_export]
+macro_rules! impl_serializable_argument_for_scalar {
+    ($inner:ty) => {
+        impl $crate::SerializableArgument for $inner {
+            fn serialize(&self) -> Result<$crate::serde_json::Value, $crate::SerializeError> {
+                use $crate::Scalar;
+                self.encode()
+            }
+        }
+    };
+}
+
+impl_serializable_argument_for_scalar!(i32);
+impl_serializable_argument_for_scalar!(f64);
+impl_serializable_argument_for_scalar!(String);
+impl_serializable_argument_for_scalar!(bool);
+impl_serializable_argument_for_scalar!(crate::Id);