+++ cynic/src/into_argument.rs
@@ -9,17 +9,18 @@ use crate::{argument::SerializableArgument, Id};
 /// There are implementations of this for most of the built in scalars to allow
 /// users to express arguments in a simple manner.  The `cynic::Enum` derive
 /// also generates impls for converting options & refs easily.
-pub trait IntoArgument<Argument>
-where
-    Argument: SerializableArgument,
-{
-    fn into_argument(self) -> Argument;
+pub trait IntoArgument<T> {
+    type Output: SerializableArgument;
+
+    fn into_argument(self) -> Self::Output;
 }
 
 impl<T> IntoArgument<T> for T
 where
     T: SerializableArgument,
 {
+    type Output = T;
+
     fn into_argument(self) -> T {
         self
     }
@@ -30,64 +31,78 @@ where
 /// Mostly just converts references to owned via cloning and
 /// non option-wrapped types into Option where appropriate.
 #[macro_export]
-macro_rules! define_into_argument_for_scalar {
+macro_rules! impl_into_argument_for_options {
     ($inner:ty) => {
         impl $crate::IntoArgument<Option<$inner>> for $inner {
+            type Output = Option<$inner>;
+
             fn into_argument(self) -> Option<$inner> {
                 Some(self)
             }
         }
 
-        impl $crate::IntoArgument<Option<$inner>> for &$inner {
-            fn into_argument(self) -> Option<$inner> {
-                Some(self.clone())
+        impl<'a> $crate::IntoArgument<Option<$inner>> for &'a $inner {
+            type Output = Option<&'a $inner>;
+
+            fn into_argument(self) -> Option<&'a $inner> {
+                Some(self)
             }
         }
     };
 }
 
-macro_rules! define_into_argument_for_option_refs {
+macro_rules! impl_into_argument_for_option_refs {
     ($inner:ty) => {
-        impl $crate::IntoArgument<Option<$inner>> for Option<&$inner> {
-            fn into_argument(self) -> Option<$inner> {
-                self.cloned()
+        impl<'a> $crate::IntoArgument<Option<$inner>> for Option<&'a $inner> {
+            type Output = Option<&'a $inner>;
+
+            fn into_argument(self) -> Option<&'a $inner> {
+                self
             }
         }
 
-        impl $crate::IntoArgument<Option<$inner>> for &Option<$inner> {
-            fn into_argument(self) -> Option<$inner> {
-                self.clone()
+        impl<'a> $crate::IntoArgument<Option<$inner>> for &'a Option<$inner> {
+            type Output = Option<&'a $inner>;
+
+            fn into_argument(self) -> Option<&'a $inner> {
+                self.as_ref()
             }
         }
     };
 }
 
-define_into_argument_for_scalar!(i32);
-define_into_argument_for_scalar!(f64);
-define_into_argument_for_scalar!(String);
-define_into_argument_for_scalar!(bool);
-define_into_argument_for_scalar!(Id);
+impl_into_argument_for_options!(i32);
+impl_into_argument_for_options!(f64);
+impl_into_argument_for_options!(String);
+impl_into_argument_for_options!(bool);
+impl_into_argument_for_options!(Id);
 
-define_into_argument_for_option_refs!(i32);
-define_into_argument_for_option_refs!(f64);
-define_into_argument_for_option_refs!(String);
-define_into_argument_for_option_refs!(bool);
-define_into_argument_for_option_refs!(Id);
+impl_into_argument_for_option_refs!(i32);
+impl_into_argument_for_option_refs!(f64);
+impl_into_argument_for_option_refs!(String);
+impl_into_argument_for_option_refs!(bool);
+impl_into_argument_for_option_refs!(Id);
 
-impl IntoArgument<String> for &str {
-    fn into_argument(self) -> String {
-        self.to_string()
+impl<'a> IntoArgument<String> for &'a str {
+    type Output = &'a str;
+
+    fn into_argument(self) -> &'a str {
+        self
     }
 }
 
-impl IntoArgument<Option<String>> for &str {
-    fn into_argument(self) -> Option<String> {
-        Some(self.to_string())
+impl<'a> IntoArgument<Option<String>> for &'a str {
+    type Output = Option<&'a str>;
+
+    fn into_argument(self) -> Option<&'a str> {
+        Some(self)
     }
 }
 
-impl IntoArgument<Option<String>> for Option<&str> {
-    fn into_argument(self) -> Option<String> {
-        self.map(|s| s.to_string())
+impl<'a> IntoArgument<Option<String>> for Option<&'a str> {
+    type Output = Option<&'a str>;
+
+    fn into_argument(self) -> Option<&'a str> {
+        self
     }
 }