+++ cynic/src/selection_set.rs
@@ -672,13 +672,13 @@ mod tests {
                 let mut args = vec![Argument::new(
                     "required_arg",
                     "String!",
-                    serde_json::Value::String(required.required_arg),
+                    required.required_arg,
                 )];
                 if optionals.opt_string.is_some() {
                     args.push(Argument::new(
                         "opt_string",
                         "String",
-                        serde_json::Value::String(optionals.opt_string.unwrap()),
+                        optionals.opt_string.unwrap(),
                     ));
                 }
                 field("nested", args, fields)